removing layer of choice type service - not compiling
